[IOTMID-220] Adding support for perseo configuration
This commit adds suport for perseo configuration when managing
flows.

It will transform the input flow (or flows) into a series of
rule creation requests and send them to perseo-fe server. It
will also remove any rule when a particular flow is deleted.

In order to do that it was needed to encapsulate all flow
boxes into an attribute in the input JSON description (attribute
"boxes").

Furthermore, this commit adds a new 'config.js' file that holds
all the initial configuration - iotagent-json style.

Change-Id: I6f960d400bcfd0d2db9e134ca1285d97810b1915